Adapter refactor
